+++ src/components/Bracket/index.ts
@@ -1,0 +1,4 @@
+export { default as OrigamBracket } from './OrigamBracket.vue'
+export { default as OrigamBracketRound } from './OrigamBracketRound.vue'
+export { default as OrigamBracketMatch } from './OrigamBracketMatch.vue'
+export { default as OrigamBracketCompetitor } from './OrigamBracketCompetitor.vue'